backend/requirements.txt
# Django Core
Django>=5.0,<6.1
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-filter>=24.0

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
social-auth-app-django>=5.4.0
python-social-auth>=0.3.6

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Security
python-decouple>=3.8
argon2-cffi>=23.1.0

# Task Queue & Scheduling
celery>=5.3.0
django-celery-beat>=2.5.0
redis>=5.0.0

# Email
django-anymail>=10.2

# File Storage
django-storages>=1.14.0
boto3>=1.34.0
Pillow>=10.2.0

# API Documentation
drf-spectacular>=0.27.0

# Utilities
python-dateutil>=2.8.2
gunicorn>=21.2.0
whitenoise>=6.6.0

# AI Features (Optional)
openai>=1.12.0

# Export Features
reportlab>=4.1.0
openpyxl>=3.1.2

# Development
python-dotenv>=1.0.0
